VERSION 1.0.0 bugfix and incompatible change to -q
This version is now under the CLL (which means it is PD).

Some bugs are fixed:

-i now can be used with producers, too.
If you catch STDERR, then STDIN and STDERR swap positions.
Use -d if you do not like that.

Option -q now has '' as default and no more '\n'.  Should be less confusing and is the better default.

Please note that this is compatible to previous often used fix "-p ''" (which now can be left away).
If you want to see continuation lines to show an LF, then please add -p $'\n'!

The version was bumped to 1.0.0 according to http://semver.org
as the change of -q is not really backward compatible.