VERSION 2.1.0 option -i0 catches stdin
$ unbuffered | consumer

can now be written as

$ unbuffered -i0 consumer

Before -i0 tried to catch output of consumers on FD 0,
which makes no sense as input should be input only.

(If you really need to catch output of producers to
stdout, use IO-redirection of stdin to some free FD,
and catch that.  Note that in that case, input on
stdin is not supported.  This is nothing new.)

Only stdin is handles special, all other FDs are
catched on the output side of forked programs.

Why is unbufferd catching output of forked programs?

$ something | consumer

gives you the exit code of consumer.  But

$ something | consumer | unbuffered

would catch the exit code of unbuffered
(you can 'set -o pipefail', but that's not the same).

To get the exit code of consumer, unbuffered
must be able to get hands on the exit code of
consumers.  Hence it only can be written as:

$ something | unbuffered producer

The only exception to this now is -i0.