VERSION 2.2.0: Options -k and -w added
Option -k allows to kill unbuffered when the command terminates.
This mainly is needed when the FD does not close due to this
termination, like with daemons or if wrapping stdin.
When -i0 is not in effect this might lose some data,
as unbuffered might terminate before all data is transferred.

Option -w changes the FD for where the -i FD is sent to.
Usually this is stdout.  For example -i2 -w2 reads from
stderr and outputs to stderr.  In that case you should use
option -o3 or similar, to change the dump output, else both
goes to stderr for -w2.  This is not prohibited.